requirements.txt
# requirements.txt

# Core
fastapi
uvicorn
streamlit

# LLM & embeddings
langchain
langchain-community
langchain-openai
langchain-chroma
langchain-ollama
chromadb
ollama

# Document parsing & OCR
unstructured[all-docs]

# Utilities
python-dotenv
pillow
pyyaml

pymupdf
langchain-text-splitters